局部微调

+++ src/main/resources/static/js/community.js
@@ -1,40 +1,50 @@
 
-/* TODO 前端校验回复内容是否为空
- *提交一级回复 */
-function pComment() {
+/* 提交一级回复 */
+function pComment(e) {
     var questionId = $("#question_id").val();
     var content = $("#comment_content").val();
-    $.ajax({
-        type:"POST",
-        url:"/comment",
-        contentType:'application/json',
-        data:JSON.stringify({
-            "pid":1,
-            "content":content,
-            "level":1,
-            "questionId":questionId
-        }),
-        success:function (response) {
-            if(response.code === 200){
-                $("#comment_section").hide();
-            }
-            else{
-                if(response.code === 4){
-                    var isAccepted = confirm(response.message);
-                    if(isAccepted) {
-                        window.open("https://github.com/login/oauth/authorize?client_id=08b076ae3cf904cf324c&redirect_uri=http://localhost:8080/callback&scope=user&state=1&&allow_signup=true");
-                        window.localStorage.setItem("closable","true");
+    var reg = /^\s*$/;
+    debugger;
+    if(content != null && !reg.test(content)){
+        $.ajax({
+            type:"POST",
+            url:"/comment",
+            contentType:'application/json',
+            data:JSON.stringify({
+                "pid":1,
+                "content":content,
+                "level":1,
+                "questionId":questionId
+            }),
+            success:function (response) {
+                if(response.code === 200){
+                    $("#comment_section").hide();
+                }
+                else{
+                    if(response.code === 4){
+                        var isAccepted = confirm(response.message);
+                        if(isAccepted) {
+                            window.open("https://github.com/login/oauth/authorize?client_id=08b076ae3cf904cf324c&redirect_uri=http://localhost:8080/callback&scope=user&state=1&&allow_signup=true");
+                            window.localStorage.setItem("closable","true");
+                        }
+                    }
+                    else {
+                        alert(response.message);
                     }
                 }
-                else {
-                    alert(response.message);
-                }
-            }
-            console.log(response);
-        },
-        dataType:"json"
-    });
+                console.log(response);
+            },
+            dataType:"json"
+        });
+    }
+    else {
+        //TODO 回复内容为空的时候
+        alert("回复内容不可为空");
+
+    }
 }
+
+
 
 function comment() {
